requirements.txt
asgiref==3.8.1
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==5.1.3
django-allauth==65.2.0
django-channels==0.7.0
django-environ==0.11.2
humanize==4.11.0
idna==3.10
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.0.0
psycopg2==2.9.10
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.2
urllib3==2.3.0
